+++ new features/electricty.py
